src/agent/requirements.txt
livekit-agents>=0.10.0
livekit-plugins-openai>=0.10.0
livekit-plugins-elevenlabs>=0.8.0
livekit-plugins-silero>=0.6.0
python-dotenv>=1.0.0
aiohttp>=3.11.0
openai>=1.57.0
elevenlabs>=1.17.0
websockets>=14.0
mcp>=1.3.2
httpx>=0.28.0
pydantic>=2.10.0